started work on correct_confidence for crossentropy

+++ src/loss/mod.rs
@@ -3,5 +3,5 @@ use crate::FynnArray;
 pub mod categorical_crossentropy;
 
 pub trait Loss {
-    fn calculate(y_pred: FynnArray, y_true: Vec<f64>); 
+    fn calculate(y_pred: FynnArray, y_true: Vec<usize>); 
 }
